requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
celery[redis]==5.3.1
redis==4.6.0
requests==2.31.0
python-dotenv==1.0.0
feedgen==0.9.0
